LayoutTests/fast/dom/Geolocation/exposed-geolocation-interfaces-expected.txt

Checks that the Geolocation interfaces are exposed and that their name is correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS windowProperties.indexOf('Geolocation') >= 0 is true
PASS windowProperties.indexOf('GeolocationCoordinates') >= 0 is true
PASS windowProperties.indexOf('GeolocationPosition') >= 0 is true
PASS windowProperties.indexOf('GeolocationPositionError') >= 0 is true
PASS successfullyParsed is true

TEST COMPLETE
